formatting and ignores

+++ rust/src/mmu/mbc1.rs
@@ -44,9 +44,9 @@ impl Mmu {
         if self.ram_enable {
             match self.ram_size {
                 0x00 => (),
-                0x02 => self.extern_ram[(loc - 0xA000) as usize] = val,
+                0x02 => self.extern_ram[loc as usize - 0xA000] = val,
                 0x03 => {
-                    self.extern_ram[(loc - 0xA000 + 0x2000 * self.ram_bank as u16) as usize] = val
+                    self.extern_ram[loc as usize - 0xA000 + 0x2000 * self.ram_bank as usize] = val
                 }
                 _ => (),
             }
